Extract and share `capture_cur_position` and `Position`
